test: split test_worktree_manager into per-theme modules (Refs #2879)
