+++ tests/unit/worktree_manager/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for agent/worktree_manager.py, split by theme (#2879)."""
